requirements.txt
amqp==2.5.2
billiard==3.6.3.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==7.0
Flask==1.0.4
idna==2.9
intervaltree==3.0.2
itsdangerous==1.1.0
Jinja2==2.9.6
kombu==4.6.8
MarkupSafe==1.0
numpy==1.18.1
pandas==1.0.1
pep8==1.7.0
python-dateutil==2.8.1
pytz==2019.3
raven==6.2.1
requests==2.23.0
six==1.14.0
sortedcontainers==2.1.0
SQLAlchemy==1.3.13
urllib3==1.25.8
vine==1.3.0
Werkzeug==1.0.0
celery==4.4.0
redis==3.4.1
gunicorn==20.0.4
requests-oauthlib==1.3.0
oauthlib==0.7.2
pycryptodome==3.9.7